Move StoreFile to same package as StoreType
