+++ community/kernel/src/test/java/org/neo4j/kernel/impl/factory/CommunityEditionModuleTest.java
@@ -25,7 +25,7 @@ import java.util.function.Predicate;
 
 import org.neo4j.kernel.impl.index.IndexConfigStore;
 import org.neo4j.kernel.impl.store.MetaDataStore;
-import org.neo4j.kernel.impl.storemigration.StoreFile;
+import org.neo4j.kernel.impl.store.StoreFile;
 import org.neo4j.kernel.impl.storemigration.StoreFileType;
 import org.neo4j.kernel.impl.transaction.log.PhysicalLogFile;
